Added input-main block with new grouping
