--- blocks/test.json
@@ -1,5 +1,0 @@
-{
-  "type": "test_block",
-  "message0": "Test Block",
-  "colour": 160
-}
