+++ blocks/train-setup.json
@@ -1,0 +1,21 @@
+{
+    "type": "train_setup",
+    "message0": "Train with batch size %1 using %2 as target",
+    "args0": [
+        {
+            "type": "field_number",
+            "name": "batch_size",
+            "value": 32,
+            "min": 0
+        },
+        {
+            "type": "field_input",
+            "name": "target_column",
+            "text": "\"column\""
+        }
+    ],
+    "colour":160,
+    "tooltip": "Provide info about training setup",
+    "nextStatement": null,
+    "previousStatement": null
+}
