+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "@blockly/field-grid-dropdown": "^5.0.7",
     "babel-loader": "^9.1.3",
     "blockly": "^11.1.1",
     "express": "^4.19.2",
